# openapi.yaml
openapi: 3.0.0
info:
  title: Swagger Petstore - OpenAPI 3.1
  description: "The api docs duh"
  version: 1.0.0
servers:
  - url: http://localhost:3000
...
paths:
...
  /offer/list/{id}:
    get:
      tags:
        - offer
      summary: Gets all open offers for a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Offer"
  /offer/{id}:
    post:
      tags:
        - offer
      summary: Claim offer parts
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: amount
          in: query
          schema:
            type: integer
      responses:
        200:
          description: Success
  /offer:
    put:
      tags:
        - offer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                units:
                  type: integer
                categoryId:
                  type: string
                product:
                  type: string
                from:
                  type: string
                  description: "Date string"
                  example: "iso-date-string"
                to:
                  type: string
                  description: "Date string"
                  example: "iso-date-string"

      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: number
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Offer:
      type: object
      properties:
        id:
          type: string
        ownerId:
          type: string
        units:
          type: integer
        categoryId:
          type: string
        product:
          type: string
        from:
          type: string
          description: "Date string"
          example: "iso-date-string"
        to:
          type: string
          description: "Date string"
          example: "iso-date-string"
security:
  - bearerAuth: []